Create olympics.csv
120 Years of Olympic History

This dataset provides athlete and event data for all Olympics held from 1896 to 2016. 

This dataset was compiled by Kaggle user Randi H. Griffin. For more information about the dataset and ideas for how to use it, visit https://www.kaggle.com/heesoo37/120-years-of-olympic-history-athletes-and-results/home. 

When using this dataset, note that the Winter and Summer Games were held in the same year up until 1992. After that, they began to be staggered, such that the Summer and Winter games alternate on a four-year cycle. A common mistake people make when analyzing this data is to assume that the Summer and Winter Games have always been staggered.
 

Columns in this dataset are as follows:
 
 
ID
An ID number assigned to this athlete based on their sequential order in the dataset 

Name
The name of the athlete

Sex
The gender of the athlete

Age
The age of the athlete

Height
The height of the athlete, in centimeters

Weight
The weight of the athlete, in kilograms

Team
The country this athlete represents

NOC
The three-letter abbreviation for the country represented by the athlete

Games
The year and season for this Olympic event

City
The city in which these Olympics was held

Sport
The sport of the event

Event
The name of the event

Medal
The medal won by the athlete ("NA" if no medal was won)